--- client/src/components/organisms/Form/types.ts
@@ -1,20 +1,0 @@
-export type InputType = 'text'
-
-interface FormOptionType {
-  label: string
-  value: string
-}
-
-export interface OptionType {
-  readOnly?: boolean
-  multiline?: boolean
-  affix?: string
-  helperText?: string
-  width?: number
-  selected?: FormOptionType
-  options?: Array<FormOptionType>
-  noOptionsMessage?: string
-  isMulti?: boolean
-  isDynamic?: boolean
-  copy?: boolean
-}
